release(npm): move platforms to wrapper/
